Include dates, write to file, and some other options
Liked the work you did, and thank you for doing the heavy lifting!

I made a change because I wanted to use a comma as a delimiter rather than a space, and then realized that others might want to, so I made it a variable that users could edit.

Then I decided that it would be useful to write the data to a file that I could import straight into a spreadsheet, so I created an option to output to screen as originally written, or to a text file (added variable for file name), or to both. In either of those cases, it will append data to the existing file if it exists, or create on if it doesn't. Finally, I thought it would be helpful to include the date data, so I added that in as well, but again, made it optional. 

As far as I can tell, I have the defaults set so that out of the box the version with my edits works exactly the same as yours. There are just more options now.

I am thinking about running this monthly as a cron job. I just need to figure out what time of day I want it to run (probably 00:01 GMT), and whether I want to figure out how to make it create a unique file name.

Thanks again for doing this!